compiles and runs correctly with new hermite

+++ src/integrals/gaussianintegrals.cpp
@@ -197,14 +197,22 @@ inline double GaussianIntegrals::coulomb2D(const unsigned int& i, const unsigned
     /* calculate full coulomb integral in 2D case */
 
     // grab hermite coefficients
-    const std::vector<int>& HCIx = HC(GaussianBasis::Cartesian::getn(i,0));
-    const std::vector<int>& HCIy = HC(GaussianBasis::Cartesian::getn(i,1));
-    const std::vector<int>& HCJx = HC(GaussianBasis::Cartesian::getn(j,0));
-    const std::vector<int>& HCJy = HC(GaussianBasis::Cartesian::getn(j,1));
-    const std::vector<int>& HCKx = HC(GaussianBasis::Cartesian::getn(k,0));
-    const std::vector<int>& HCKy = HC(GaussianBasis::Cartesian::getn(k,1));
-    const std::vector<int>& HCLx = HC(GaussianBasis::Cartesian::getn(l,0));
-    const std::vector<int>& HCLy = HC(GaussianBasis::Cartesian::getn(l,1));
+    const std::vector<long int>& HCIx =
+        HC(GaussianBasis::Cartesian::getn(i,0));
+    const std::vector<long int>& HCIy =
+        HC(GaussianBasis::Cartesian::getn(i,1));
+    const std::vector<long int>& HCJx =
+        HC(GaussianBasis::Cartesian::getn(j,0));
+    const std::vector<long int>& HCJy =
+        HC(GaussianBasis::Cartesian::getn(j,1));
+    const std::vector<long int>& HCKx =
+        HC(GaussianBasis::Cartesian::getn(k,0));
+    const std::vector<long int>& HCKy =
+        HC(GaussianBasis::Cartesian::getn(k,1));
+    const std::vector<long int>& HCLx =
+        HC(GaussianBasis::Cartesian::getn(l,0));
+    const std::vector<long int>& HCLy =
+        HC(GaussianBasis::Cartesian::getn(l,1));
 
     // find the maximum index (upper limit for coeffs and integrals needed)
     unsigned int pmax = Methods::max(HCIx.size(), HCKx.size(), HCJx.size(),
@@ -273,18 +281,30 @@ inline double GaussianIntegrals::coulombElement3D(const unsigned int& ix, const
 inline double GaussianIntegrals::coulomb3D(const unsigned int& i, const unsigned
         int& j, const unsigned int& k, const unsigned int& l) {
     /* calculate full coulomb integral in 2D case */
-    const std::vector<int>& HCIx = HC(GaussianBasis::Cartesian::getn(i,0));
-    const std::vector<int>& HCIy = HC(GaussianBasis::Cartesian::getn(i,1));
-    const std::vector<int>& HCIz = HC(GaussianBasis::Cartesian::getn(i,2));
-    const std::vector<int>& HCKx = HC(GaussianBasis::Cartesian::getn(k,0));
-    const std::vector<int>& HCKy = HC(GaussianBasis::Cartesian::getn(k,1));
-    const std::vector<int>& HCKz = HC(GaussianBasis::Cartesian::getn(k,2));
-    const std::vector<int>& HCJx = HC(GaussianBasis::Cartesian::getn(j,0));
-    const std::vector<int>& HCJy = HC(GaussianBasis::Cartesian::getn(j,1));
-    const std::vector<int>& HCJz = HC(GaussianBasis::Cartesian::getn(j,2));
-    const std::vector<int>& HCLx = HC(GaussianBasis::Cartesian::getn(l,0));
-    const std::vector<int>& HCLy = HC(GaussianBasis::Cartesian::getn(l,1));
-    const std::vector<int>& HCLz = HC(GaussianBasis::Cartesian::getn(l,2));
+    const std::vector<long int>& HCIx =
+        HC(GaussianBasis::Cartesian::getn(i,0));
+    const std::vector<long int>& HCIy =
+        HC(GaussianBasis::Cartesian::getn(i,1));
+    const std::vector<long int>& HCIz =
+        HC(GaussianBasis::Cartesian::getn(i,2));
+    const std::vector<long int>& HCKx =
+        HC(GaussianBasis::Cartesian::getn(k,0));
+    const std::vector<long int>& HCKy =
+        HC(GaussianBasis::Cartesian::getn(k,1));
+    const std::vector<long int>& HCKz =
+        HC(GaussianBasis::Cartesian::getn(k,2));
+    const std::vector<long int>& HCJx =
+        HC(GaussianBasis::Cartesian::getn(j,0));
+    const std::vector<long int>& HCJy =
+        HC(GaussianBasis::Cartesian::getn(j,1));
+    const std::vector<long int>& HCJz =
+        HC(GaussianBasis::Cartesian::getn(j,2));
+    const std::vector<long int>& HCLx =
+        HC(GaussianBasis::Cartesian::getn(l,0));
+    const std::vector<long int>& HCLy =
+        HC(GaussianBasis::Cartesian::getn(l,1));
+    const std::vector<long int>& HCLz =
+        HC(GaussianBasis::Cartesian::getn(l,2));
     
     unsigned int pmax = Methods::max(HCIx.size(), HCKx.size(), HCJx.size(),
             HCLx.size(), HCIy.size(), HCKy.size(), HCJy.size(), HCLy.size(),
